progress

+++ src/components/ThreeWindow/shaders/textFragment.js
@@ -1,4 +1,4 @@
-const fragShader =  `precision highp float;
+const fragShader =  `precision mediump float;
           #define M_PI 3.14159265359
 
           // Varying
